Separate provider class
